Renamed Controller to Workspace
kept an empty file called __Controller__ with a wrapper class that inherits from Workspace.  The code works just fine, but the word 'controller' needs to be rooted out.
