+++ OpenPNM/Base/__init__.py
@@ -20,10 +20,12 @@ r"""
 
 """
 import logging as logging
+from .__Workspace__ import Workspace
 from .__Controller__ import Controller
 from .__ModelsDict__ import ModelsDict
 from . import __Tools__ as Tools
 from .__Core__ import Core
+
 
 # Set up logging to file - see previous section for more details
 log_format = \
